Rewrite solver completely
Major commit! Since the first implementation of the solver was very
rushed and, somewhat correlatedly, of exceedingly poor code quality, I
decided to go back and rewrite it essentially completely. There's still
some work left to be done, but I think the result is already much more
presentable than what we had previously.

A big part of this rework is shifting most of the work from `solver.cpp`
to `toroidal_list.cpp`. The solver was degeneralized to only be able to
solve 9x9 sudokus in exchange for being able to inline and slim down the
matrix setup routine. The internal solver interface was lifted to a
more modern design.

Apart from the changes to the solver, a minimal test is also included,
as well as some cleanup in `utility.hpp`.

Finally, this commit comprises extensions to the CMake files which will
now build the project with native optimizations and LTO/IPO in release
mode if such are available.